Use jupytext to pair coordinate-frames.ipynb with .py (nomarker) script
Open up Jupyter Notebook and use the new jupytext menu options
File --> Jupytext --> Pair Notebook with nomarker Script to pair
those two files (.ipynb and .py). Then use File --> Save and
Checkpoint and File --> Close and Halt to close the file.

This updates the .ipynb *and* .md file to include some additional
jupytext specific metadata and updates the .py file as well. When
compared with the .py file available with the jupytext light script
format, it's identical with one exception:
 - Remove all simple # + and # - tag pairs