Enclose metadata at top of .doxygen.md in <!-- --- ---- -->
"Since the metadata is in YAML format, I'd prefer to keep the
--- ---, and enclose it in the HTML comment <!-- --->, rather
than replace it with the HTML comment"

Note that there needs to be a blank whitespace line between
<!-- and --- and again between --- and --> for Doxygen to parse
and render this correctly. Without those whitespace lines,
Doxygen still captures the metadata and tries to include it in
the generated .html pages.